Rename InviteParticipantsModal to ManageParticipantsModal
